requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
Flask-Cors==4.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7
boto3==1.28.40
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.2.0
Pillow==10.0.1
email-validator==2.0.0
PyJWT==2.8.0
